check if brew installed; added cmdline tools

+++ setup/setup.sh
@@ -1,16 +1,23 @@
 source ~/.config/danielbaird-osx/bin/bash-color-vars.sh
 
+PREFIX="${BLUE}◈◈◈◈◈ setup ◈◈◈◈◈  "
 function declare {
-  echo "${BLUE}◈◈◈setup◈◈◈  ${YELLOW}$1${RESTORE}"
+    # tell the user something
+    echo "${PREFIX} ${YELLOW}$1${RESTORE}"
 }
 function demand {
-  read -p "${BLUE}◈◈◈setup◈◈◈  ${YELLOW}► $1${RESTORE}" -n 1 -r
+  # ask the user for a y/n to progress
+    read -p "${PREFIX} ${YELLOW}► $1${RESTORE}" -n 1 -r
 }
 function reassure {
-  echo "${BLUE}◈◈◈setup◈◈◈  ${YELLOW}${GREEN}✓${YELLOW} $1${RESTORE}"
-  echo
+    # tell the user something, with a tick at the start
+    echo "${PREFIX} ${YELLOW}${GREEN}✓${YELLOW} $1${RESTORE}"
+    echo
 }
-
+function commandexists {
+    # check if there is a command already
+    command -v "$1" >/dev/null 2>&1
+}
 # link in ~/.bash-profile -----------------------------------------------------
 if [ -e ~/.bash_profile ]; then
     demand "Remove existing .bash_profile (y/n)? "
@@ -49,12 +56,22 @@ done
 popd > /dev/null
 
 # install homebrew ------------------------------------------------------------
-declare "Installing the Xcode command line tools requies some GUI interaction."
-declare "Click 'Install' on the forthcoming alert box."
-declare "If you see '...already installed' then you're fine."
-xcode-select --install
-/usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
-reassure "install homebrew"
+if commandexists brew; then
+    brew update
+    reassure "update homebrew"
+else
+    if commandexists gcc; then
+        reassure "Xcode commandline tools already intalled"
+    else
+        declare "Installing the Xcode command line tools requies some GUI interaction."
+        declare "Click 'Install' on the forthcoming alert box."
+        declare "If you see '...already installed' then you're fine."
+        xcode-select --install
+        reassure "install Xcode commandline tools"
+    fi
+    /usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
+    reassure "install homebrew"
+fi
 
 # run all the scripts in setup/scripts ----------------------------------------
 pushd ~/.config/danielbaird-osx/setup/scripts > /dev/null
@@ -62,7 +79,7 @@ for SCRIPT in *.sh; do
     if [ -f $SCRIPT -a -x $SCRIPT ]; then
         "$HOME/.config/danielbaird-osx/setup/scripts/$SCRIPT"
         echo
-        reassure "${SCRIPT//-/ }" | sed 's/\.sh$//'
+        reassure "`echo \"${SCRIPT//-/ }\" | sed 's/\.sh$//'`"
     fi
 done
 popd > /dev/null
